Moved sqlalchemy db implementation
