--- services/assistant/app/database/__init__.py
@@ -1,3 +1,0 @@
-from app.database.orm import get_db_factory
-
-__all__ = ["get_db_factory"]
