--- services/assistant/app/database/repositories/__init__.py
@@ -1,5 +1,0 @@
-from app.database.repositories.booking_repository import SQLBookingRepository
-from app.database.repositories.user_repository import SQLUserRepository
-from app.database.repositories.xray_repository import SQLXrayRepository
-
-__all__ = ["SQLBookingRepository", "SQLUserRepository", "SQLXrayRepository"]
