+++ services/assistant/app/database/sql_db/repositories/base_repository.py
