+++ services/assistant/app/database/sql_db/repositories/user_repository.py
@@ -3,12 +3,12 @@ from sqlalchemy.orm import InstrumentedAttribute
 
 from app.repositories import UserRepository
 
-from app.database.repositories.base_repository import SQLBaseRepository
+from app.database.sql_db.repositories.base_repository import SQLBaseRepository
 from app.schemas.user_schemas import UserCreateSchema, UserSchema, UserUpdateSchema
 
-from app.database import converters
-from app.database.models import User
-from app.database.orm import selectinload_all
+from app.database.sql_db import converters
+from app.database.sql_db.models import User
+from app.database.sql_db.orm import selectinload_all
 
 
 class SQLUserRepository(UserRepository, SQLBaseRepository):
